+++ mopidy/utils/network.py
@@ -290,6 +290,7 @@ class LineProtocol(ThreadingActor):
 
     def __init__(self, connection):
         self.connection = connection
+        self.prevent_timeout = False
         self.recv_buffer = ''
 
         if self.delimeter:
@@ -326,7 +327,8 @@ class LineProtocol(ThreadingActor):
             if line is not None:
                 self.on_line_received(line)
 
-        self.connection.enable_timeout()
+        if not self.prevent_timeout:
+            self.connection.enable_timeout()
 
     def on_stop(self):
         """Ensure that cleanup when actor stops."""